+++ services/users/src/main/scala/api/dto/request/UpdateUserRequestDTO.scala
@@ -1,0 +1,14 @@
+package api.dto.request
+
+import io.circe.Decoder
+import io.circe.generic.semiauto._
+
+case class UpdateUserRequestDTO(
+    accountDTO: Option[UpdateAccountDTO] = None,
+    profileDTO: Option[UpdateProfileDTO] = None
+)
+
+object UpdateUserRequestDTO:
+  import UpdateAccountDTO.given
+  import UpdateProfileDTO.given
+  given Decoder[UpdateUserRequestDTO] = deriveDecoder