+++ services/users/src/main/scala/api/dto/response/AvatarResponseDTO.scala
@@ -1,0 +1,6 @@
+package api.dto.response
+
+case class AvatarResponseDTO(
+    userId: String,
+    avatarUrl: String
+)