+++ services/users/src/main/scala/api/utils/RequestParser.scala
@@ -1,0 +1,9 @@
+package api.utils
+
+import cask.Request
+import io.circe.Decoder
+import io.circe.parser.decode
+
+object RequestParser:
+  def parseRequestBody[T: Decoder](request: Request): Either[String, T] =
+    decode[T](request.text()).left.map(_ => "Invalid JSON")