+++ services/users/src/main/scala/app/Main.scala
@@ -12,7 +12,12 @@ object Main extends cask.MainRoutes {
   override def host: String                = "0.0.0.0"
   override def defaultHandler: HttpHandler = new CorsHandler(super.defaultHandler)
   override def allRoutes: Seq[cask.Routes] =
-    Seq(new AuthRoutes(authService, userService), new UserRoutes(userService), new AppRoutes())
+
+    Seq(
+      new AuthRoutes(authService, userService, eventPublisher),
+      new UserRoutes(authService, userService, userQueryService, eventPublisher),
+      new AppRoutes()
+    )
 
   sys.addShutdownHook {
     try {