+++ services/users/src/main/scala/http/HttpSecurity.scala
@@ -1,0 +1,13 @@
+package http
+
+import cask.Request
+import infrastructure.keycloak.KeycloakJwtVerifier.authorizeUser
+import infrastructure.keycloak.KeycloakJwtVerifier.verifyToken
+
+object HttpSecurity:
+  def authenticateAndAuthorize(request: Request, userId: String): Either[String, Unit] =
+    for
+      userAccessToken <- AuthHeaderExtractor.extractBearer(request)
+      payload         <- verifyToken(userAccessToken)
+      _               <- authorizeUser(payload, userId)
+    yield ()